# testing/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CyberSec Assistant API",
    "description": "Backend API for Cyber Security Assistant",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/system/api-usage": {
      "get": {
        "tags": [
          "admin",
          "admin",
          "system"
        ],
        "summary": "Get Api Usage",
        "description": "Get API usage statistics for external services.\n\nRequires admin or security analyst role.",
        "operationId": "get_api_usage_api_admin_system_api_usage_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 1,
              "default": 30,
              "title": "Days"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/system/cache": {
      "get": {
        "tags": [
          "admin",
          "admin",
          "system"
        ],
        "summary": "Get Cve Cache Stats",
        "description": "Get CVE lookup cache statistics.\n\nRequires admin or security analyst role.",
        "operationId": "get_cve_cache_stats_api_admin_system_cache_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/system/cache/{cve_id}": {
      "delete": {
        "tags": [
          "admin",
          "admin",
          "system"
        ],
        "summary": "Clear Cve Cache",
        "description": "Clear specific CVE cache entry.\n\nRequires admin role.",
        "operationId": "clear_cve_cache_api_admin_system_cache__cve_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cve_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cve Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin",
          "admin",
          "system"
        ],
        "summary": "Refresh Cve Cache",
        "description": "Force refresh specific CVE cache entry.\n\nRequires admin role.",
        "operationId": "refresh_cve_cache_api_admin_system_cache__cve_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cve_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Cve Id"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CacheRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/system/analytics": {
      "get": {
        "tags": [
          "admin",
          "admin",
          "system"
        ],
        "summary": "Get System Analytics",
        "description": "Get system-wide analytics and statistics.\n\nRequires admin or security analyst role.",
        "operationId": "get_system_analytics_api_admin_system_analytics_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemAnalyticsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/system/dashboard/cached": {
      "get": {
        "tags": [
          "admin",
          "admin",
          "system"
        ],
        "summary": "Get Cached Dashboard Stats",
        "description": "Get cached dashboard statistics for improved performance.\n\nUses Redis caching with 10-minute TTL to reduce database load.\nSet force_refresh=true to bypass cache and recompute statistics.\n\nRequires admin or security analyst role.",
        "operationId": "get_cached_dashboard_stats_api_admin_system_dashboard_cached_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "force_refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force Refresh"
            }
          },
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/system/dashboard/clear-cache": {
      "post": {
        "tags": [
          "admin",
          "admin",
          "system"
        ],
        "summary": "Clear Dashboard Cache",
        "description": "Clear dashboard statistics cache.\n\nForces recomputation on next dashboard load.\n\nRequires admin role.",
        "operationId": "clear_dashboard_cache_api_admin_system_dashboard_clear_cache_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/system/cache/stats": {
      "get": {
        "tags": [
          "admin",
          "admin",
          "system"
        ],
        "summary": "Get System Cache Stats",
        "description": "Get Redis cache statistics and health.\n\nRequires admin or security analyst role.",
        "operationId": "get_system_cache_stats_api_admin_system_cache_stats_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-User-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CacheRefreshRequest": {
        "properties": {
          "force_refresh": {
            "type": "boolean",
            "title": "Force Refresh",
            "default": true
          }
        },
        "type": "object",
        "title": "CacheRefreshRequest",
        "description": "Request model for refreshing CVE cache"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SystemAnalyticsResponse": {
        "properties": {
          "total_users": {
            "type": "integer",
            "title": "Total Users"
          },
          "active_users": {
            "type": "integer",
            "title": "Active Users"
          },
          "total_scans": {
            "type": "integer",
            "title": "Total Scans"
          },
          "total_news": {
            "type": "integer",
            "title": "Total News"
          },
          "severity_distribution": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Severity Distribution"
          },
          "api_usage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Api Usage"
          }
        },
        "type": "object",
        "required": [
          "total_users",
          "active_users",
          "total_scans",
          "total_news",
          "severity_distribution",
          "api_usage"
        ],
        "title": "SystemAnalyticsResponse",
        "description": "Response model for system analytics"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}